Chnaged argument size
As per the `op_construct_operand` function documentation, the x64
and x86 `offset` will take `size_t` or 64-bit sized offsets instead
of the previous `uint8_t` sized.

+++ libjas/instruction.c
@@ -237,7 +237,7 @@ instruction_t instr_gen(enum instructions instr, uint8_t operand_count, ...) {
   for (uint8_t i = 0; i < operand_count; i++)
     operands[i] = op_construct_operand(
         va_arg(args, enum operands),
-        va_arg(args, uint8_t), va_arg(args, void *));
+        va_arg(args, size_t), va_arg(args, void *));
 
   return (instruction_t){
       .instr = instr,
